refactor(api): extract shared software version/upload helpers to a service module
